refactor: exception 한 곳으로 모으기
