+++ src/main/java/com/dinosaur/foodbowl/global/error/ExceptionAdvice.java
@@ -1,10 +1,17 @@
 package com.dinosaur.foodbowl.global.error;
 
+import jakarta.validation.ConstraintViolation;
+import jakarta.validation.ConstraintViolationException;
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindException;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @RestControllerAdvice
@@ -23,6 +30,16 @@ public class ExceptionAdvice {
         e.getMessage());
     return ResponseEntity.status(e.getHttpStatus())
         .body(ErrorResponse.from(errorMessage));
+  }
+
+  @ExceptionHandler
+  @ResponseStatus(HttpStatus.BAD_REQUEST)
+  public Map validationExceptionHandler(ConstraintViolationException e) {
+    List<String> errors = e.getConstraintViolations()
+        .stream()
+        .map(ConstraintViolation::getMessage)
+        .collect(Collectors.toList());
+    return Collections.singletonMap("errors", errors);
   }
 
   private static String getErrorMessage(BindException e) {
